// mic-front/src/app/interfaces/mic/associate/associate.ts
export interface Associate {
  id: number | undefined
  active: boolean | undefined
  name: string | undefined
  document: string | undefined
  card_number: string | undefined
  gender: string | undefined
  birthdate: Date | undefined
  cellphone_number: string | undefined
  phone_number: string | undefined
  marital_status: string | undefined
  zip_code: string | undefined
  street: string | undefined
  number: string | undefined
  complement: string | undefined
  neighborhood: string | undefined
  city: string | undefined
  state: string | undefined
}